+++ src/string-util_test.cc
@@ -38,18 +38,18 @@ TEST(StringUtilTest, HasPrefix) {
 
 TEST(StringUtilTest, SplitString) {
     std::vector<StringPiece> result = SplitString("foo", ",");
-    EXPECT_EQ(1, result.size());
+    EXPECT_EQ(1, (int)result.size());
     EXPECT_EQ(StringPiece("foo"), result[0]);
 
     result = SplitString(",hello, world", ",");
-    EXPECT_EQ(3, result.size());
+    EXPECT_EQ(3, (int)result.size());
     EXPECT_EQ(StringPiece(""), result[0]);
     EXPECT_EQ(StringPiece("hello"), result[1]);
     EXPECT_EQ(StringPiece(" world"), result[2]);
 
     // Also test with trailing, empty field
     result = SplitString(",hello, world,", ",");
-    EXPECT_EQ(4, result.size());
+    EXPECT_EQ(4, (int)result.size());
     EXPECT_EQ(StringPiece(""), result[0]);
     EXPECT_EQ(StringPiece("hello"), result[1]);
     EXPECT_EQ(StringPiece(" world"), result[2]);
